+++ app/src/main/java/com/sopt/now/compose/data/api/service/ReqresService.kt
@@ -1,6 +1,6 @@
 package com.sopt.now.compose.data.api.service
 
-import com.sopt.now.compose.data.dto.response.ResponseReqresDto
+import com.sopt.now.compose.data.api.dto.response.ReqresResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
@@ -9,5 +9,5 @@ interface ReqresService {
     @GET("api/users")
     suspend fun getUserList(
         @Query("page") page: Int
-    ): Response<ResponseReqresDto>
+    ): Response<ReqresResponse>
 }
