+++ app/src/main/java/com/sopt/now/compose/feature/auth/LoginViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewModelScope
 import com.sopt.now.compose.ui.core.view.UiState
 import com.sopt.now.compose.data.api.ApiFactory
 import com.sopt.now.compose.data.api.UserIdProvider
-import com.sopt.now.compose.data.dto.request.RequestLoginDto
+import com.sopt.now.compose.data.api.dto.request.RequestLoginDto
 import com.sopt.now.compose.feature.util.KeyStorage.ID_DEFAULT
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
