+++ app/src/main/java/com/sopt/now/compose/feature/auth/SignUpViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.sopt.now.compose.ui.core.view.UiState
 import com.sopt.now.compose.data.api.ApiFactory
-import com.sopt.now.compose.data.dto.request.RequestSignUpDto
-import com.sopt.now.compose.feature.model.User
+import com.sopt.now.compose.data.api.dto.request.RequestSignUpDto
+import com.sopt.now.compose.model.User
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
