+++ app/src/main/java/com/sopt/now/compose/model/HomeSealedItem.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose.feature.model
+package com.sopt.now.compose.model
 
 import android.os.Parcelable
 import androidx.annotation.DrawableRes
