+++ app/src/main/java/com/sopt/now/compose/model/User.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose.feature.model
+package com.sopt.now.compose.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
